rename translation files
This fixes case-insensitive file name conflicts, e.g. editor.ts Editor.ts
