+++ src/components/RunButton.tsx
@@ -7,7 +7,7 @@ import { downloadAndRun } from '../actions/hub';
 import { RootState } from '../reducers';
 import { HubRuntimeState } from '../reducers/hub';
 import ActionButton, { ActionButtonProps } from './ActionButton';
-import { TooltipId } from './button';
+import { TooltipId } from './button-i18n';
 import runIcon from './images/run.svg';
 
 type StateProps = Pick<ActionButtonProps, 'enabled'>;
