+++ src/components/button-i18n.test.ts
@@ -2,8 +2,8 @@
 // Copyright (c) 2020 The Pybricks Authors
 
 import { lookup } from '../../test';
-import { TooltipId } from './button';
-import en from './button.en.json';
+import { TooltipId } from './button-i18n';
+import en from './button-i18n.en.json';
 
 describe('Ensure .json file has matches for TooltipIds', () => {
     test.each(Object.values(TooltipId))('%s', (id) => {
