+++ src/components/notification-i18n.test.ts
@@ -2,8 +2,8 @@
 // Copyright (c) 2020 The Pybricks Authors
 
 import { lookup } from '../../test';
-import { MessageId } from '../reducers/notification';
-import en from './notification.en.json';
+import { MessageId } from './notification-i18n';
+import en from './notification-i18n.en.json';
 
 describe('Ensure .json file has matches for MessageIds', () => {
     test.each(Object.values(MessageId))('%s', (id) => {
