+++ src/components/notification-i18n.ts
@@ -1,0 +1,16 @@
+// SPDX-License-Identifier: MIT
+// Copyright (c) 2020 The Pybricks Authors
+// File: components/notification-i18n.ts
+// Notification translation keys.
+
+export enum MessageId {
+    BleCannotWriteWithoutResponse = 'ble.cannotWriteWithoutResponse',
+    BleConnectFailed = 'ble.connectFailed',
+    BleGattPermission = 'ble.gattPermission',
+    BleGattServiceNotFound = 'ble.gattServiceNotFound',
+    BleNoWebBluetooth = 'ble.noWebBluetooth',
+    ProgramChanged = 'editor.programChanged',
+    ServiceWorkerSuccess = 'serviceWorker.success',
+    ServiceWorkerUpdate = 'serviceWorker.update',
+    YesReloadProgram = 'editor.yesReloadProgram',
+}
