+++ src/components/terminal-i18n.en.json
