+++ src/components/terminal-i18n.test.ts
@@ -2,8 +2,8 @@
 // Copyright (c) 2020 The Pybricks Authors
 
 import { lookup } from '../../test';
-import { TerminalStringId } from './terminal';
-import en from './terminal.en.json';
+import { TerminalStringId } from './terminal-i18n';
+import en from './terminal-i18n.en.json';
 
 describe('Ensure .json file has matches for TerminalStringIds', () => {
     test.each(Object.values(TerminalStringId))('%s', (id) => {
