+++ src/components/terminal-i18n.ts
@@ -1,5 +1,7 @@
 // SPDX-License-Identifier: MIT
 // Copyright (c) 2020 The Pybricks Authors
+// File: components/terminal-i18n.ts
+// Terminal translation keys.
 
 export enum TerminalStringId {
     Copy = 'terminal.copy',
